Split into util namespace and importable implicits.
